requirements.txt
# BLT Preflight Advisory Engine
# Python dependencies

# Core (no external dependencies needed for basic functionality)
# The engine is designed to work with Python standard library only

# Optional: For enhanced functionality
# requests>=2.28.0  # For future GitHub API integration
# pyyaml>=6.0       # For YAML configuration support

# --- Scout-Risk AI Advisory Engine (GSoC 2026 Enhancement) ---
fastapi>=0.115.0 # Backend for Gemini AI integration
uvicorn>=0.30.0 # ASGI server to run the AI engine
python-multipart
google-genai>=0.3.0
requests>=2.32.0  # For fetching live security documentation
beautifulsoup4>=4.12.0  # Live OWASP Cheat Sheet mapping
python-dotenv # Environment variable management for API keys
pydantic>=2.9.0
